Started refactoring User around Goals rather than Goals & GoalSlugs
